+++ tests/test_serving.py
@@ -1,4 +1,5 @@
 import os
+import re
 from pathlib import Path
 from typing import Any
 
@@ -7,6 +8,7 @@ import requests
 from slugify import slugify
 from starlette.testclient import TestClient
 
+from foxcross.constants import SLUGIFY_REGEX, SLUGIFY_REPLACE
 from foxcross.enums import MediaTypes
 from foxcross.exceptions import PostProcessingError, PredictionError, PreProcessingError
 from foxcross.serving import ModelServing, ModelServingRunner, compose_models
@@ -139,16 +141,23 @@ def test_index_single_model_serving():
 def test_predict_multi_model_serving():
     app = compose_models(__name__, debug=True)
     client = TestClient(app)
+
+    add_one_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, AddOneModel.__name__)
+    )
     add_one_response = client.post(
-        f"{slugify(AddOneModel.__name__)}/predict/",
+        f"{add_one_slugified}/predict/",
         headers={"Accept": MediaTypes.JSON.value},
         json=add_one_data,
     )
     assert add_one_response.status_code == 200
     assert add_one_response.json() == add_one_result_data
 
+    add_five_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, AddFiveModel.__name__)
+    )
     add_five_response = client.post(
-        f"{slugify(AddFiveModel.__name__)}/predict/",
+        f"{add_five_slugified}/predict/",
         headers={"Accept": MediaTypes.JSON.value},
         json=add_five_data,
     )
@@ -159,10 +168,17 @@ def test_predict_multi_model_serving():
 def test_index_multi_model_serving():
     app = compose_models(__name__, debug=True)
     client = TestClient(app)
-    add_one_response = client.get(f"{slugify(AddOneModel.__name__)}/")
+
+    add_one_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, AddOneModel.__name__)
+    )
+    add_one_response = client.get(f"{add_one_slugified}/")
     assert add_one_response.status_code == 200
 
-    add_five_response = client.get(f"{slugify(AddFiveModel.__name__)}/")
+    add_five_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, AddFiveModel.__name__)
+    )
+    add_five_response = client.get(f"{add_five_slugified}/")
     assert add_five_response.status_code == 200
 
     root_response = client.get("/")
@@ -179,16 +195,20 @@ def test_index_multi_model_serving():
 def test_endpoints_multi_model_serving(endpoint, first_expected, second_expected):
     app = compose_models(__name__, debug=True)
     client = TestClient(app)
+    add_one_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, AddOneModel.__name__)
+    )
     add_one_response = client.get(
-        f"{slugify(AddOneModel.__name__)}{endpoint}",
-        headers={"Accept": MediaTypes.JSON.value},
+        f"{add_one_slugified}{endpoint}", headers={"Accept": MediaTypes.JSON.value}
     )
     assert add_one_response.status_code == 200
     assert add_one_response.json() == first_expected
 
+    add_five_slugified = slugify(
+        re.sub(SLUGIFY_REGEX, SLUGIFY_REPLACE, AddFiveModel.__name__)
+    )
     add_five_response = client.get(
-        f"{slugify(AddFiveModel.__name__)}{endpoint}",
-        headers={"Accept": MediaTypes.JSON.value},
+        f"{add_five_slugified}{endpoint}", headers={"Accept": MediaTypes.JSON.value}
     )
     assert add_five_response.status_code == 200
     assert add_five_response.json() == second_expected