+++ foxcross/filters.py
@@ -1,0 +1,5 @@
+def has_attr(value, attr):
+    if hasattr(value, attr):
+        return True
+    else:
+        return False