+++ docs/release-notes.md
@@ -4,6 +4,7 @@
 * Renamed `BadDataFormatError` to `PredictionError`
 * Updated route name generation in `runner.ModelServingRunner.compose` to create
 better urls
+* Added bootstrap 4 to improve model frontend
 
 ## 0.6.0
 * Added `PreProcessingError` for use with `pre_process_input`